Move structs & enums to dedicated file
